refactor: rearrange App + Dashboard to separate business logic
