+++ src/App/App.js
@@ -1,37 +1,41 @@
 import React, { useEffect } from "react";
 import Dashboard from "../Dashboard";
 import Header from "./components/Header";
-import useDashboardRoute from "../Dashboard/hooks/useDashboardRoute";
 import Search from "../Search";
 import useDashboardStore from "../Dashboard/hooks/useDashboardStore";
 import useDashboardDefaults from "../Dashboard/hooks/useDashboardDefaults";
-import { Box, CircularProgress } from "@material-ui/core";
-import { QueryClient, QueryClientProvider } from "react-query";
+import useLanguageStore from "../Language/useLanguageStore";
+import { Legend } from "../Legend";
+import { Map } from "../Map";
+import Panel from "../Panel/Panel";
+import { Tooltip } from "../Tooltip";
+import Body from "./components/Body";
+import Loading from "./components/Loading";
+import {
+  getCurrentRouteParams,
+  parseRoute,
+  populateRoute,
+  ROUTE_TEMPLATE,
+  validateRoute,
+} from "./router";
+import RouteProvider from "../Router/Router";
+import useAppRouter from "./hooks/useAppRouter";
+import useRouter from "../Router/useRouter";
 
-// Create a client
-const queryClient = new QueryClient();
-
-const App = ({ config }) => {
+const App = ({ lang = "en", langDict, config }) => {
   // pull ready state from the store
   const ready = useDashboardStore((state) => state.ready);
 
-  const { route, getCurrentParams } = useDashboardRoute();
-
-  // load params from route on mount (if any)
-  const routeDefaults = getCurrentParams();
-
-  // update route on state changes
+  //👇 update language on changes
+  const setLanguage = useLanguageStore((state) => state.setLanguage);
   useEffect(() => {
-    // do not set in not ready or undefined route
-    if (!ready || route.indexOf("undefined") > -1 || route.indexOf("null") > -1)
-      return null;
-    window.history && window.history.replaceState(null, null, route);
-  }, [route, ready]);
+    setLanguage(lang, langDict);
+  }, [lang, langDict, setLanguage]);
 
-  // set the default dashboard state
+  // 👇 set the default dashboard state based on route params + config
   useDashboardDefaults({
     ...config,
-    ...routeDefaults,
+    ...getCurrentRouteParams(),
     defaultViewport: {
       zoom: config.zoom,
       latitude: config.latitude,
@@ -39,25 +43,33 @@ const App = ({ config }) => {
     },
   });
 
+  useAppRouter();
+
   return (
-    <QueryClientProvider client={queryClient}>
+    // <RouteProvider
+    //   template={ROUTE_TEMPLATE}
+    //   parse={parseRoute}
+    //   populate={populateRoute}
+    //   validator={validateRoute}
+    //   values={routeValues}
+    // >
+    <Dashboard>
       <Header>
         <Search />
       </Header>
       {ready ? (
-        <Dashboard />
+        <Body>
+          <Map>
+            <Legend />
+          </Map>
+          <Panel />
+        </Body>
       ) : (
-        <Box
-          display="flex"
-          alignItems="center"
-          justifyContent="center"
-          width="100%"
-          height="100%"
-        >
-          <CircularProgress />
-        </Box>
+        <Loading />
       )}
-    </QueryClientProvider>
+      <Tooltip />
+    </Dashboard>
+    // </RouteProvider>
   );
 };
 
@@ -141,6 +153,64 @@ App.defaultProps = {
     zoom: 8,
     latitude: 32.74,
     longitude: -96.96,
+  },
+  lang: "en",
+  langDict: {
+    en: {
+      APP_TITLE: "North Texas Evictions",
+      BUTTON_MENU_OPEN: "Menu",
+      METRIC_CPR: "Child Poverty Rate",
+      METRIC_PVR: "Poverty Rate",
+      METRIC_EF: "Eviction Filings",
+      METRIC_EFR: "Filings Per 1000 Renters",
+      METRIC_MFA: "Median Filing Amount",
+      METRIC_PCW: "% Non-Hispanic White",
+      METRIC_PCB: "% Black",
+      METRIC_PCH: "% Latinx",
+      METRIC_PCA: "% Asian",
+      METRIC_MHI: "Median Household Income",
+      METRIC_MGR: "Median Gross Rent",
+      METRIC_MPV: "Median Property Value",
+      METRIC_PRH: "% Renter Homes",
+      METRIC_RB: "Rent Burden",
+      METRIC_TFA: "Total Filing Amount",
+      REGION_COUNTIES: "Counties",
+      REGION_TRACTS: "Census Tracts",
+      REGION_ZIPS: "ZIP Codes",
+      REGION_CITIES: "Cities",
+      REGION_DISTRICTS: "Council Districts",
+      LEGEND_SUMMARY: "between {{start}} and {{end}}",
+      BUTTON_CHANGE_OPTIONS: "Change Data Options",
+      SELECT_CHOROPLETH: "Demographic Metric",
+      SELECT_BUBBLE: "Eviction Metric",
+      SELECT_REGION: "Region",
+      SELECT_DATE_START: "Start Date",
+      SELECT_DATE_END: "End Date",
+      SELECT_COURT: "Court",
+      SELECT_DATE_RANGE: "Date Range",
+      TITLE_DATA_OPTIONS: "Data Options",
+      SUMMARY: "Summary ({{dateRange}})",
+      SUMMARY_EF: "Total Eviction Filings",
+      SUMMARY_TFA: "Total Amount Filed",
+      SUMMARY_SERIES: "Filings By Day",
+      SUMMARY_UPDATED: "Data last updated on {{date}}.",
+      LEGEND: "Map Legend",
+      LEGEND_TITLE: "Currently Viewing",
+      FLAG_MFA:
+        "Median filings amounts are only available within Dallas County. ",
+      FLAG_SHORT_EFR:
+        "Filings per 1,000 renters will be small for short time ranges.",
+      HINT_TOTAL_FILINGS: "",
+      HINT_TOTAL_AMOUNT:
+        "Filing amounts are only reported within Dallas County, the actual total is much higher.",
+      LABEL_ALL_COURTS: "All Courts",
+      LABEL_FIT_BOUNDS: "Zoom to all {{region}}",
+      LABEL_UNAVAILABLE: "Unavailable",
+      LABEL_REGION_UNAVAILABLE: "Unavailable for {{region}}",
+      LABEL_SHOW_LEGEND: "Show Full Legend",
+      LABEL_HIDE_LEGEND: "Show Full Map",
+      LABEL_SHOW_DATA_OPTIONS: "Show All Data Options",
+    },
   },
 };
