+++ src/App/hooks/useMediaQueries.js
@@ -1,7 +1,13 @@
+import { useMemo } from "react";
 import { useMediaQuery } from "@material-ui/core";
 
 export default function useMediaQueries() {
   const isMobile = useMediaQuery("(max-width:600px)");
 
-  return { isMobile };
+  return useMemo(
+    () => ({
+      isMobile,
+    }),
+    [isMobile]
+  );
 }
