+++ src/Legend/Legend.js
@@ -29,6 +29,7 @@ import usePrecinctNames from "../Data/usePrecinctNames";
 import LegendRow from "./components/LegendRow";
 import useTogglePanel from "../Panel/useTogglePanel";
 import useMediaQueries from "../App/hooks/useMediaQueries";
+import useDashboardStatus from "../Dashboard/hooks/useDashboardStatus";
 
 /**
  * Returns a prefix and label for the date range text in the legend
@@ -136,6 +137,9 @@ const Legend = ({ classes, ...props }) => {
   const [activeBubble, setActiveBubble, bubbleMetrics] = useDashboardBubble();
   const [activeChoropleth, setActiveChoropleth, choroplethMetrics] =
     useDashboardChoropleth();
+
+  // useDashboardStatus();
+  console.count("Legend");
 
   // prepare language
   const langKeys = [
